+++ fotoobo/tools/fmg/main.py
@@ -8,24 +8,26 @@ from pathlib import Path
 from fotoobo.exceptions.exceptions import GeneralWarning
 from fotoobo.helpers.config import config
 from fotoobo.helpers.files import load_json_file
+from fotoobo.helpers.result import Result
 from fotoobo.inventory import Inventory
 
 log = logging.getLogger("fotoobo")
 
 
-def assign(adoms: str, policy: str, host: str, timeout: int = 60) -> None:
+def assign(adoms: str, policy: str, host: str, timeout: int = 60) -> Result[str]:
     """
     Assign the global policy to the given ADOM
 
     Args:
-        adoms (str):   The ADOMs to assign the global policy to. Specify multiple ADOMs as a comma
-            separated list (no spaces).
-        policy (str):  Specify the global policy to assign [Default: 'default'].
-        host (str):    The FortiManager defined in inventory.
-        timeout (int): Timeout in sec. to wait for the FortiManager task to finish [Default: 60].
+        adoms:  The ADOMs to assign the global policy to. Specify multiple ADOMs as a comma
+                separated list (no spaces).
+        policy: Specify the global policy to assign [Default: 'default'].
+        host:   The FortiManager defined in inventory.
+        timeout Timeout in sec. to wait for the FortiManager task to finish [Default: 60].
     """
+    result = Result[str]()
     inventory = Inventory(config.inventory_file)
-    fmg = inventory.get(host, "fortimanager")[host]
+    fmg = inventory.get_item(host, "fortimanager")
     fmg.login()
 
     log.debug("Assigning global policy/objects to ADOM %s", adoms)
@@ -40,18 +42,21 @@ def assign(adoms: str, policy: str, host: str, timeout: int = 60) -> None:
                 if message["end_tm"] > 0
                 else "unfinished"
             )
-            getattr(log, level)(
-                "FortiManager task %s: %s%s (%s)",
-                message["task_id"],
-                message["name"],
-                f" / {message['detail']}" if message["detail"] else "",
-                elapsed,
-            )
+            result_message = f"{message['task_id']}: {message['name']}"
+            if message["detail"]:
+                result_message += f" / {message['detail']}"
+
+            result_message += f" ({elapsed})"
+            getattr(log, level)(result_message)
+            result.push_message(host, result_message, level)
             if message["history"]:
                 for line in message["history"]:
-                    getattr(log, level)("- %s", line["detail"])
+                    result_message = f"- {line['detail']}"
+                    getattr(log, level)(result_message)
+                    result.push_message(host, result_message, level)
 
     fmg.logout()
+    return result
 
 
 def post(file: Path, adom: str, host: str) -> None:
@@ -70,7 +75,7 @@ def post(file: Path, adom: str, host: str) -> None:
         raise GeneralWarning(f"there is no data in the given file ({file})")
 
     inventory = Inventory(config.inventory_file)
-    fmg = inventory.get(host, "fortimanager")[host]
+    fmg = inventory.get_item(host, "fortimanager")
     fmg.login()
 
     log.debug("FortiManager post command ...")