+++ WHATSNEW.md
@@ -1,5 +1,8 @@
 ### Added
 
+- Option `--smtp` for `fmg assign`
+- Method `inventory.get_item()` to only get one single item from the inventory
+
 ### Changed
 
 ### Removed